Analytics: Add external link support (#242)
WIP: 

Adds a new component `OutboundLink`  for links that take you outside of the app, so we track when the users use it.

Thanks  @W3stside, @alfetopito for the help


I didn a small modifications to be as close as possible as the signature of the tative `<a />` element.  So now `to` is called `href`, and we don't required to add the `eventLabel` (we use the link by default)


This PR adds the ExternalLink component, in the 3 links in the footer, plus in the BlockExplorer component.